+++ contracts/PriceOracle.sol
@@ -7,10 +7,10 @@ contract PriceOracle {
     bool public constant isPriceOracle = true;
 
     /**
-      * @notice Get the underlying price of a cToken asset
-      * @param cToken The cToken to get the underlying price of
-      * @return The underlying asset price mantissa (scaled by 1e18).
-      *  Zero means the price is unavailable.
-      */
-    function getUnderlyingPrice(CToken cToken) external view returns (uint);
+     * @notice Get the underlying price of a cToken asset
+     * @param cToken The cToken to get the underlying price of
+     * @return The underlying asset price mantissa (scaled by 1e18).
+     *  Zero means the price is unavailable.
+     */
+    function getUnderlyingPrice(CToken cToken) external view returns (uint256);
 }
